Allow float duration in EFFECT

+++ ScriptedEvents/Actions/Player/EffectAction.cs
@@ -37,7 +37,7 @@
             new Argument("players", typeof(Player[]), "The players to affect.", true),
             new Argument("effect", typeof(EffectType), "The effect to give or remove.", true),
             new Argument("intensity", typeof(int), "The intensity of the effect, between 0-255. Variables are supported. Defaults to 1.", false),
-            new Argument("duration", typeof(int), "The duration of the effect, or no duration for a permanent effect. Variables are supported.", false),
+            new Argument("duration", typeof(float), "The duration of the effect, or no duration for a permanent effect. Variables are supported.", false),
         };
 
         /// <inheritdoc/>
@@ -56,11 +56,11 @@
                     return new(MessageType.LessThanZeroNumber, this, "intensity", Arguments[3]);
             }
 
-            int duration = 0;
+            float duration = 0;
 
             if (Arguments.Length > 4)
             {
-                duration = (int)Arguments[4];
+                duration = (float)Arguments[4];
                 if (duration < 0)
                     return new(MessageType.LessThanZeroNumber, this, "duration", Arguments[4]);
             }
